Finalize unified line emission, replace old version
